requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.8.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.35.0
PyPDF2>=3.0.0
reportlab>=4.0.0
requests>=2.31.0
